+++ client/admin.py
@@ -3,6 +3,11 @@ from django.contrib import admin
 from .models import *
 
 
+
+class User(admin.StackedInline):
+    admin.site.register(CV)
+
+
 class Subtasks_Inline(admin.StackedInline):
 
     model = SubTasks
