--- client/templates/client/client_edit_experience.html
@@ -1,103 +1,0 @@
-{% extends 'client/main_template_client.html' %}
-{% load staticfiles %}
-
-{% block title %}
-    Edit Experience
-{% endblock %}
-
-{% block content %}
-
-    <!-- JS добавления однотипных форм. ONLY FOR THIS CUSTOM FORM -->
-    <script type="text/javascript">
-        $(document).ready(function () {
-            var count = 1;
-            var origin = $("#origin_div").clone();
-
-            $("#add_more").on("click", function () {
-                var add = origin.clone();
-                var current_div = add.children('div');
-
-                for (var i = 0; i < 6; i++) {
-                    if (i === 1) {  // 1 select
-                        var element_select = $(current_div).children('select');
-                        console.log(element_select);
-                        var attr_name_s = $(element_select).attr('name');
-                        console.log(attr_name_s);
-                        $(element_select).attr('name', attr_name_s + count);
-                    } else {    // 5 inputs
-                        var element_input = $(current_div[i]).children('input');
-                        console.log(element_input);
-                        var attr_name_i = $(element_input).attr('name');
-                        console.log(attr_name_i);
-                        $(element_input).attr('name', attr_name_i + count);
-                    }
-                }
-                count++;
-                console.log(count);
-                add.appendTo("#div_to_add_new");
-            });
-        });
-    </script>
-
-    <div class="row">
-        <div class="col-md-6">
-            <form method="post">
-                {% csrf_token %}
-
-                <div id="div_to_add_new">
-                    <div id="origin_div">
-                        <h3>Experience 1</h3>
-
-                        <div class="form-group">
-                            <label for="id_experience_1">Название организации</label>
-                            <input type="text" class="form-control" id="id_experience_1" name="experience_1">
-                        </div>
-
-                        <div class="form-group">
-                            <label for="id_experience_2">Область работы (сфера)</label>
-                            <select multiple class="form-control" id="id_experience_2" name="experience_2">
-                                <option value="1">1</option>
-                                <option value="2">2</option>
-                                <option value="3">3</option>
-                                <option value="4">4</option>
-                                <option value="5">5</option>
-                            </select>
-                        </div>
-
-                        <div class="form-group">
-                            <label for="id_experience_3">Должность</label>
-                            <input type="text" class="form-control" id="id_experience_3" name="experience_3">
-                        </div>
-
-                        <h5>Период работы:</h5>
-
-                        <div class="form-group">
-                            <label for="id_exp_date_start">Дата начала</label>
-                            <input type="date" class="form-control" id="id_exp_date_start" name="exp_date_start">
-                        </div>
-
-                        <div class="form-group">
-                            <label for="id_exp_date_end">Дата окончания</label>
-                            <input type="date" class="form-control" id="id_exp_date_end" name="exp_date_end">
-                        </div>
-
-                        <div class="form-group">
-                            <label for="id_experience_4">Обязанности на рабочем месте</label>
-                            <input type="text" class="form-control" id="id_experience_4" name="experience_4">
-                        </div>
-
-                    </div>
-                </div>
-
-                <input type="button" value="Добавить опыт работы" id="add_more"
-                       class="btn btn-outline-primary m-sm-1">
-
-                <button type="submit" class="btn btn-outline-success m-sm-1"
-                        id="save">Соранить все данные
-                </button>
-            </form>
-
-        </div>
-    </div>
-
-{% endblock %}
