--- client/templates/client/client_edit_skills.html
@@ -1,48 +1,0 @@
-{% extends 'client/main_template_client.html' %}
-{% load staticfiles %}
-
-{% block title %}
-    Edit Skills
-{% endblock %}
-
-{% block content %}
-    <!-- JS добавления однотипных форм -->
-    <script type="text/javascript">
-        $(document).ready(function () {
-            var origin = $("#id_skill").clone();
-            var count = 1;
-            $("#add_more").on("click", function () {
-                var add = origin.clone();
-                count += 1;
-                add.attr('value', '');
-                add.attr("id", "new_skill");
-                add.appendTo("#div_to_add_new");
-            });
-        });
-    </script>
-
-    <form method="post" id="origin_form">
-        {% csrf_token %}
-        <div id="div_to_add_new">
-
-            <h3 class="text-center">Навыки</h3>
-            {% if not data.cl_skill %}
-                <input type="text" class="form-control my-2" id="id_skill" name="skill">
-            {% else %}
-                {% for s in data.cl_skill %}
-                    <input type="text" class="form-control my-2" id="id_skill" name="skill"
-                           value="{{ s }}">
-                {% endfor %}
-            {% endif %}
-
-        </div>
-
-        <div class="text-center">
-            <a class="btn btn-outline-warning m-2" href="{% url 'client_edit' %}">Назад</a>
-            <button type="submit" class="btn btn-outline-success m-2" id="save">Соранить все данные</button>
-            <input type="button" value="Добавить Навыки" id="add_more" class="btn btn-outline-primary m-2">
-        </div>
-
-    </form>
-
-{% endblock %}
