+++ client/templates/client/edit_forms/client_edit_skills.html
@@ -1,0 +1,67 @@
+{% extends 'client/main_template_client.html' %}
+{% load staticfiles %}
+
+{% block title %}
+    Edit Skills
+{% endblock %}
+
+{% block content %}
+    <!-- JS добавления однотипных форм -->
+    <script type="text/javascript">
+        $(document).ready(function () {
+            var origin = $("#id_skill").clone();
+            var count = 1;
+
+            $("#add_more").on("click", add);
+            $("#id_skill").on("change", data_l);
+
+            function add() {
+                var add = origin.clone().on("change", data_l);
+                count += 1;
+                add.attr('value', '');
+                add.attr("id", "new_skill");
+                add.appendTo("#div_to_add_new");
+            }
+
+            function data_l() {
+                var us_in = $(this).val();
+                console.log(us_in);
+                $.ajax({
+                    'url': "{% url 'skills_data_list' %}",
+                    'data': {'us_in': us_in},
+                    success: function (data) {
+                        var json_data = jQuery.parseJSON(data);
+                        console.log(json_data.words);
+                        {#//$("#data_list_2")#}
+                    }
+                });
+            }
+        });
+    </script>
+
+    <form method="post" id="origin_form">
+        {% csrf_token %}
+        <div id="div_to_add_new">
+            <h3 class="text-center">Навыки</h3>
+            {% for s in data.cl_skill %}
+                <input type="text" class="form-control my-2" id="id_skill"
+                       name="skill"
+                       value="{{ s }}" list="data_list_1">
+                <datalist id="data_list_1">
+                    <option id="dl_option" value=""></option>
+                </datalist>
+            {% empty %}
+                <input type="text" class="form-control my-2" id="id_skill"
+                       name="skill" list="data_list_2">
+                <datalist id="data_list_2">
+                    <option id="dl_option" value=""></option>
+                </datalist>
+            {% endfor %}
+        </div>
+        <div class="text-center">
+            <a class="btn btn-outline-warning m-2" href="{% url 'client_edit' %}">Назад</a>
+            <button type="submit" class="btn btn-outline-success m-2" id="save">Соранить все данные</button>
+            <input type="button" value="Добавить Навыки" id="add_more" class="btn btn-outline-primary m-2">
+        </div>
+    </form>
+{% endblock %}
