+++ client/templates/client/edit_forms/cv_empty.html
@@ -1,0 +1,36 @@
+<div id="origin_div">
+    <h3 id="cv_h">Резюме</h3>
+    <div class="form-group">
+        <label for="id_position">Должность</label>
+        <input type="text" class="form-control" id="id_position" name="position">
+    </div>
+    <div class="form-group">
+        <label for="id_employment">Занятость</label>
+        <select class="form-control" id="id_employment" name="employment">
+            {% for emp in data.employment %}
+            <option value="{{ emp }}">{{ emp }}</option>
+            {% endfor %}
+        </select>
+    </div>
+    <div class="form-group">
+        <label for="id_time_job">График работы</label>
+        <select class="form-control" id="id_time_job" name="time_job">
+            {% for tj in data.time_job %}
+            <option selected value="{{ tj }}">{{ tj }}</option>
+            {% endfor %}
+        </select>
+    </div>
+    <h5>Уровень заработной платы</h5>
+    <div class="form-group">
+        <label for="id_salary">Сумма</label>
+        <input type="text" class="form-control" id="id_salary" name="salary">
+    </div>
+    <div class="form-group">
+        <label for="id_type_salary">Валюта</label>
+        <select class="form-control" id="id_type_salary" name="type_salary">
+            {% for ts in data.type_salary %}
+            <option value="{{ ts }}">{{ ts }}</option>
+            {% endfor %}
+        </select>
+    </div>
+</div>
