+++ client/templates/client/edit_forms/experience_empty.html
@@ -1,0 +1,34 @@
+<div id="origin_div">
+    <h3>Experience 1</h3>
+    <div class="form-group">
+        <label for="id_experience_1">Название организации</label>
+        <input type="text" class="form-control" id="id_experience_1" name="experience_1">
+    </div>
+    <div class="form-group">
+        <label for="id_experience_2">Область работы (сфера)</label>
+        <select multiple class="form-control" id="id_experience_2" name="experience_2">
+            <option value="1">1</option>
+            <option value="2">2</option>
+            <option value="3">3</option>
+            <option value="4">4</option>
+            <option value="5">5</option>
+        </select>
+    </div>
+    <div class="form-group">
+        <label for="id_experience_3">Должность</label>
+        <input type="text" class="form-control" id="id_experience_3" name="experience_3">
+    </div>
+    <h5>Период работы:</h5>
+    <div class="form-group">
+        <label for="id_exp_date_start">Дата начала</label>
+        <input type="date" class="form-control" id="id_exp_date_start" name="exp_date_start">
+    </div>
+    <div class="form-group">
+        <label for="id_exp_date_end">Дата окончания</label>
+        <input type="date" class="form-control" id="id_exp_date_end" name="exp_date_end">
+    </div>
+    <div class="form-group">
+        <label for="id_experience_4">Обязанности на рабочем месте</label>
+        <input type="text" class="form-control" id="id_experience_4" name="experience_4">
+    </div>
+</div>
