--- client/templates/client/education_empty.html
@@ -1,53 +1,0 @@
-{% spaceless %}
-    <div id="div_to_add_new">
-        <div id="origin_div">
-
-            <h3 id="edu_h">Образование</h3>
-
-            <div class="form-group">
-                <label for="id_institution">Учебное заведение</label>
-                <input type="text" class="form-control" id="id_institution" name="institution">
-            </div>
-            <div class="form-group">
-                <label for="id_subject_area">Предметная область</label>
-                <input type="text" class="form-control" id="id_subject_area" name="subject_area">
-            </div>
-            <div class="form-group">
-                <label for="id_specialization">Специальность</label>
-                <input type="text" class="form-control" id="id_specialization" name="specialization">
-            </div>
-            <div class="form-group">
-                <label for="id_qualification">Квалификация</label>
-                <input type="text" class="form-control" id="id_qualification" name="qualification">
-            </div>
-            <p>Период обучения:</p>
-            <div class="form-group">
-                <label for="id_date_start">Дата начала</label>
-                <input type="date" class="form-control" id="id_date_start" name="date_start">
-            </div>
-            <div class="form-group">
-                <label for="id_date_end">Дата окончания</label>
-                <input type="date" class="form-control" id="id_date_end" name="date_end">
-            </div>
-
-            {#                    <div id="cert_to_add">#}
-            {#                        <h3 id="cert_h">Сертификаты</h3>#}
-            {#                        <div id="cert_origin">#}
-            <div class="form-group">
-                <label for="id_certificate_img">Сертификат (img)</label>
-                <input type="file" class="form-control" id="id_certificate_img" name="certificate_img">
-            </div>
-
-            <div class="form-group">
-                <label for="id_certificate_url">Сертификат (ссылка)</label>
-                <input type="url" class="form-control" id="id_certificate_url" name="certificate_url">
-            </div>
-            {#                        </div>#}
-            <!-- копироваться будет сюда -->
-            {#                    </div>#}
-            {#                    <input type="button" value="Добавить сертификат" id="add_more_cert"#}
-            {#                           class="btn btn-outline-primary m-2">#}
-
-        </div>
-    </div>
-{% endspaceless %}
