+++ client/views.py
@@ -1,8 +1,3 @@
-import logging
-from collections import defaultdict
-from time import perf_counter
-
-#from PIL import Image
 from django.contrib import auth
 from django.shortcuts import redirect, render, get_object_or_404
 from django.template.context_processors import csrf
@@ -18,29 +13,27 @@ from .models import Vacancy, Resume
 from .forms import UploadImgForm, AddSkillForm, AddSkillFormSet, OpinionForm, AnswerForm, MessageForm
 
 from django.views.generic import View, TemplateView
+from django.views.generic import TemplateView
+
+from client.edit.check_clients import (client_check, load_client_img)
+from client.edit.edit_forms import (EducationFormSet, CertificateForm, SabClassFormSet, UploadImgForm)
+from client.edit.pages_get import (edit_page_get, skills_page_get, cv_page_get, education_page_get, experience_page_get)
+from client.edit.pages_post import (skills_page_post, edit_page_post, photo_page_post, form_edu_post,
+                                    education_page_post, cv_page_post, experience_page_post)
+from client.forms import (OpinionForm, AnswerForm, MessageForm)
+from client.models import *
 
 
-from BelHardCRM.settings import MEDIA_URL
-from client.work_with_db import (load_client_img, load_edit_page, client_check, load_skills_page, load_education_page,
-                                 load_cv_edition_page)
-from .forms import OpinionForm, AnswerForm, MessageForm
-from .forms import UploadImgForm, EducationFormSet, CertificateFormSet
-from .models import *
-from .utility import (check_input_str, check_home_number, check_telegram, check_phone, pars_cv_request,
-                      pars_edu_request, pars_exp_request)
-
-
-def client_main_page(request):  #!!!!!!!!!!!!!!!!!!!!!Alert 
+def client_main_page(request):  # !!!!!!!!!!!!!!!!!!!!!Alert
     response = csrf(request)
 
-    readtask = len(Tasks.objects.filter(user = request.user, readtask=False))
+    readtask = len(Tasks.objects.filter(user=request.user, readtask=False))
     chat = Chat.objects.get(members=request.user)
     unread_messages = len(Message.objects.filter(chat=chat, is_read=False).exclude(author=request.user))
     settings = Settings.objects.get(user=request.user)
     context = {'unread_messages': unread_messages, 'readtask': readtask, 'settings': settings}
 
-   
-    #Poland
+    # Poland
     resumes = Resume.objects.all()
     suggestions = 0
     for resume in resumes:
@@ -53,321 +46,113 @@ def client_main_page(request):  #!!!!!!!!!!!!!!!!!!!!!Alert
     return render(request=request, template_name='client/main_template_client.html', context=context)
 
 
-
 def client_profile(request):
-    response = csrf(request)
     client_instance = client_check(request.user)
-    response['client_img'] = load_client_img(client_instance)
-
+    response = {'client_img': load_client_img(client_instance),
+                }
     return render(request=request, template_name='client/client_profile.html', context=response)
 
 
-def client_edit_main(request):
-    response = csrf(request)
+# TeamRome
+class ClientEditMain(TemplateView):
+    template_name = 'client/edit_forms/client_edit_main.html'
 
-    client_instance = client_check(request.user)
+    def get(self, request, *args, **kwargs):
+        client_instance = client_check(request.user)
+        response = {'client_img': load_client_img(client_instance),
+                    'data': edit_page_get(client_instance),
+                    }
+        return render(request=request, template_name=self.template_name, context=response)
 
-    if request.method == 'POST':
-        time_0 = perf_counter()
-        print('client_edit_main - request.POST')
-
-        """ Входные данные для сохранения: """
-        user = request.user
-        user_name = check_input_str(request.POST['client_first_name'])
-        last_name = check_input_str(request.POST['client_last_name'])
-        patronymic = check_input_str(request.POST['client_middle_name'])
-        sex = Sex.objects.get(sex_word=request.POST['sex']) if request.POST['sex'] else None
-        date = request.POST['date_born'] if request.POST['date_born'] else None
-        citizenship = Citizenship.objects.get(country_word=request.POST['citizenship']) if request.POST[
-            'citizenship'] else None
-        family_state = FamilyState.objects.get(state_word=request.POST['family_state']) if request.POST[
-            'family_state'] else None
-        children = Children.objects.get(children_word=request.POST['children']) if request.POST['children'] else None
-        country = Citizenship.objects.get(country_word=request.POST['country']) if request.POST['country'] else None
-        city = City.objects.get(city_word=request.POST['city']) if request.POST['city'] else None
-        street = check_input_str(request.POST['street'])
-        house = check_home_number(request.POST['house'])
-        flat = check_home_number(request.POST['flat'])
-        telegram_link = check_telegram(request.POST['telegram_link'])
-        skype = check_input_str(request.POST['skype_id'])
-        email = request.POST['email']
-        link_linkedin = request.POST['link_linkedin']
-        state = State.objects.get(state_word=request.POST['state']) if request.POST['state'] else None
-
-        print(user_name, last_name, patronymic, sex, date, citizenship, family_state, children, country, city,
-              street, house, flat, telegram_link, skype, email, link_linkedin, state)
-
-        if not client_instance:
-            """ Если карточки нету - создаём. """
-            print('User Profile DO NOT exists - creating!')
-
-            client = Client(
-                user_client=user,
-                name=user_name,
-                last_name=last_name,
-                patronymic=patronymic,
-                sex=sex,
-                date_born=date,
-                citizenship=citizenship,
-                family_state=family_state,
-                children=children,
-                country=country,
-                city=city,
-                street=street,
-                house=house,
-                flat=flat,
-                telegram_link=telegram_link,
-                skype=skype,
-                email=email,
-                link_linkedin=link_linkedin,
-                state=state,
-            )
-            client.save()
-        else:
-            """ Если карточка есть - достаём из БД Объект = Клиент_id.
-            Перезаписываем (изменяем) существующие данныев. """
-            print('User Profile exists - Overwriting user data')
-
-            client = client_instance
-            client.name = user_name
-            client.last_name = last_name
-            client.patronymic = patronymic
-            client.sex = sex
-            client.date_born = date
-            client.citizenship = citizenship
-            client.family_state = family_state
-            client.children = children
-            client.country = country
-            client.city = city
-            client.street = street
-            client.house = house
-            client.flat = flat
-            client.telegram_link = telegram_link
-            client.skype = skype
-            client.email = email
-            client.link_linkedin = link_linkedin
-            client.state = state
-            client.save()
-
-        """ Сохранение телефонных номеров клиента """
-        tel = request.POST.getlist('phone')
-        print("tel: %s" % tel)
-        for t in tel:
-            t = check_phone(t)
-            if t:
-                phone = Telephone(
-                    client_phone=client,
-                    telephone_number=t
-                )
-                phone.save()
-
-        print('client_edit_main - OK; TIME: %s' % (perf_counter() - time_0))
+    def post(self, request):
+        client_instance = client_check(request.user)
+        edit_page_post(client_instance, request)
         return redirect(to='/client/profile')
-    else:
-        print('client_edit_main - request.GET')
-
-        """ Загрузка из БД списков для выбора и данных клиента"""
-        response['client_img'] = load_client_img(client_instance)
-        response['data'] = load_edit_page(client_instance)
-
-    return render(request=request, template_name='client/client_edit_main.html', context=response)
 
 
-def client_edit_skills(request):
-    response = csrf(request)
+# TeamRome
+class ClientEditSkills(TemplateView):
+    template_name = 'client/edit_forms/client_edit_skills.html'
 
-    client_instance = client_check(request.user)
+    def get(self, request, *args, **kwargs):
+        client_instance = client_check(request.user)
+        response = {'client_img': load_client_img(client_instance),
+                    'data': skills_page_get(client_instance),
+                    }
+        return render(request=request, template_name=self.template_name, context=response)
 
-    if request.method == 'POST':
-        print("client_edit_skills - request.POST")
-
-        skills_arr = request.POST.getlist('skill') if request.POST.getlist('skill') else None
-        print("skill: %s" % skills_arr)
-
-        if any(skills_arr):
-            for s in skills_arr:
-                if s:
-                    """ ОБЪЕДИНЕНИЕ модуля Навыки с конкретным залогиненым клиентом!!! """
-                    skill = Skills(
-                        client_skills=client_instance,
-                        skill=check_input_str(s)
-                    )
-                    skill.save()
-        else:
-            print("No skills")
-
+    def post(self, request):
+        client_instance = client_check(request.user)
+        skills_page_post(client_instance, request)
         return redirect(to='/client/edit')
-    else:
-        print('client_edit_skills - request.GET')
-        response['client_img'] = load_client_img(client_instance)
-        response['data'] = load_skills_page(client_instance)
-
-    return render(request=request, template_name='client/client_edit_skills.html', context=response)
 
 
-def client_edit_photo(request):
-    response = csrf(request)
-    client_instance = client_check(request.user)
+# TeamRome
+class ClientEditPhoto(TemplateView):
+    template_name = 'client/edit_forms/client_edit_photo.html'
 
-    if request.method == 'POST':
-        print('client_edit_photo - request.POST')
+    def get(self, request, *args, **kwargs):
+        client_instance = client_check(request.user)
+        response = {'client_img': load_client_img(client_instance),
+                    'form': UploadImgForm(),
+                    }
+        return render(request=request, template_name=self.template_name, context=response)
 
-        form = UploadImgForm(request.POST, request.FILES)
-        if form.is_valid():
-            img = form.cleaned_data.get('img')
-            client_instance.img = img
-            client_instance.save()
-            """
-            в БД сохраняется УНИКАЛЬНОЕ имя картинки (пр. user_2_EntrmQR.png)
-            в папке MEDIA_URL = '/media/'
-            """
-            print('client save photo - OK')
-            return redirect(to='/client/edit')
-    else:
-        print('client_edit_photo - request.GET')
-        response['client_img'] = load_client_img(client_instance)
-        response['form'] = UploadImgForm()
-
-    return render(request=request, template_name='client/client_edit_photo.html', context=response)
-
-
-def client_edit_cv(request):
-    response = csrf(request)
-    client_instance = client_check(request.user)
-
-    if request.method == 'POST':
-        print('client_edit_cv - request.POST')
-
-        arr_cv = pars_cv_request(request.POST)
-        for cvs in arr_cv:
-            position = cvs['position']
-            employment = Employment.objects.get(employment=request.POST['employment'])
-            time_job = TimeJob.objects.get(time_job_word=request.POST['time_job'])
-            salary = cvs['salary']
-            type_salary = TypeSalary.objects.get(type_word=request.POST['type_salary'])
-
-            if any([position, employment, time_job, salary, type_salary]):
-                cv = CV(
-                    client_cv=client_instance,
-                    position=position,
-                    employment=employment,
-                    time_job=time_job,
-                    salary=salary,
-                    type_salary=type_salary,
-                )
-                cv.save()
-
-                print("CV Form - OK\n", position, employment, time_job, salary, type_salary)
-            else:
-                print('Cv form is Empty')
+    def post(self, request):
+        client_instance = client_check(request.user)
+        photo_page_post(client_instance, request)
         return redirect(to='/client/edit')
-    else:
-        print('client_edit_cv - request.GET')
-        response['client_img'] = load_client_img(client_instance)
-        response['data'] = load_cv_edition_page(client_instance)
-
-    return render(request, 'client/client_edit_cv.html', response)
 
 
-def client_edit_education(request):
-    response = csrf(request)
-    client_instance = client_check(request.user)
+# TeamRome
+class ClientEditCv(TemplateView):
+    template_name = 'client/edit_forms/client_edit_cv.html'
 
-    if request.method == 'POST':
-        print("save_client_education - request.POST")
+    def get(self, request, *args, **kwargs):
+        client_instance = client_check(request.user)
+        response = {'client_img': load_client_img(client_instance),
+                    'data': cv_page_get(client_instance),
+                    }
+        return render(request, self.template_name, response)
 
-        arr_edu = pars_edu_request(request.POST, request.FILES)
-        for edus in arr_edu:
-            institution = edus['institution']
-            subject_area = edus['subject_area']
-            specialization = edus['specialization']
-            qualification = edus['qualification']
-            date_start = edus['date_start']
-            date_end = edus['date_end']
-            link = edus['certificate_url']
-            img = edus['certificate_img']
+    def post(self, request):
+        client_instance = client_check(request.user)
+        cv_page_post(client_instance, request)
+        return redirect(to='/client/edit')
 
-            if any([institution, subject_area, specialization, qualification,
-                    date_start, date_end, img, link]):
 
-                education = Education(
-                    client_edu=client_instance,
-                    institution=institution,
-                    subject_area=subject_area,
-                    specialization=specialization,
-                    qualification=qualification,
-                    date_start=date_start if date_start else None,
-                    date_end=date_end if date_end else None
-                )
-                education.save()
+# TeamRome
+class ClientEditEducation(TemplateView):
+    template_name = 'client/edit_forms/client_edit_education.html'
 
-                certificate = Certificate(
-                    education=education,
-                    img=img,
-                    link=link
-                )
-                certificate.save()
+    def get(self, request, *args, **kwargs):
+        client_instance = client_check(request.user)
+        response = {'client_img': load_client_img(client_instance),
+                    'data': education_page_get(client_instance),
+                    }
+        return render(request, self.template_name, response)
 
-                print("Education Form - OK\n", institution, subject_area, specialization, qualification,
-                      date_start if date_start else None, date_end if date_end else None, img, link)
-            else:
-                print('Education Form is Empty')
-
+    def post(self, request):
+        client_instance = client_check(request.user)
+        education_page_post(client_instance, request)
         return redirect('/client/edit')
-    else:
-        print('client_edit_education - request.GET')
-        response['client_img'] = load_client_img(client_instance)
-        response['data'] = load_education_page(client_instance)
-
-    return render(request, 'client/client_edit_education.html', response)
 
 
-def client_edit_experience(request):
-    response = csrf(request)
-    response['client_img'] = load_client_img(request.user)
+# TeamRome
+class ClientEditExperience(TemplateView):
+    template_name = 'client/edit_forms/client_edit_experience.html'
 
-    if request.method == 'POST':
-        print("save_client_edit_experience - request POST")
+    def get(self, request, *args, **kwargs):
+        client_instance = client_check(request.user)
+        response = {'client_img': load_client_img(client_instance),
+                    "data": experience_page_get(client_instance),
+                    }
+        return render(request, self.template_name, response)
 
-        arr = pars_exp_request(request.POST)
-        for dic in arr:
-            organisation = dic['experience_1']
-            position = dic['experience_3']
-            start_date = dic['exp_date_start']
-            end_date = dic['exp_date_end']
-            duties = dic['experience_4']
-
-            if any([organisation, position, start_date, end_date, duties]):
-                client = Client.objects.get(user_client=request.user)
-
-                experiences = Experience(
-                    client_exp=client,
-                    name=organisation,
-                    position=position,
-                    start_date=start_date if start_date else None,
-                    end_date=end_date if end_date else None,
-                    duties=duties if duties else None
-                )
-                experiences.save()
-
-                spheres = dic['experience_2']
-                for s in spheres:
-                    if s:
-                        """ Save ManyToManyField 'sphere' """
-                        sp = Sphere(sphere_word=s)
-                        sp.save()
-                        experiences.sphere.add(sp)
-
-                print("Experience Form - OK\n", organisation, spheres, position, start_date if start_date else None,
-                      end_date if end_date else None, duties if duties else None)
-            else:
-                print('Experience Form is Empty')
-
+    def post(self, request):
+        client_instance = client_check(request.user)
+        experience_page_post(client_instance, request)
         return redirect('/client/edit')
-    else:
-        print('save_client_edit_experience - request GET')
-
-    return render(request, 'client/client_edit_experience.html', response)
 
 
 class MessagesView(View):
@@ -382,7 +167,10 @@ class MessagesView(View):
             chat = None
 
         unread_messages = len(Message.objects.filter(chat=chat, is_read=False).exclude(author=request.user))
-        context = {'user_profile': request.user, 'unread_messages': unread_messages, 'chat': chat, 'form': MessageForm()}
+        context = {'user_profile': request.user,
+                   'unread_messages': unread_messages,
+                   'chat': chat,
+                   'form': MessageForm()}
         client_instance = client_check(request.user)
         context['client_img'] = load_client_img(client_instance)
         return render(request, 'client/client_chat.html', context)
@@ -459,7 +247,6 @@ class OpinionDelete(View):
 def client_login(request):  # ввести логин/пароль -> зайти в систему
     res = csrf(request)
     res['url'] = 'login'
-    print(res)
     if request.POST:
         password = request.POST['password']
         user = request.POST['user']
@@ -497,88 +284,30 @@ def tasks(request):
     task_false = sorted(task_false, key=lambda x:x.endtime,  reverse=True)
     client_instance = client_check(request.user)
     return render(request, 'client/tasks.html', context = {'task' : task,
-                                                           'task_false': task_false,
+                                                        'task_false': task_false,
                                                            'client_img':load_client_img(client_instance)})
 
 
+# TeamRome
 class FormEducation(TemplateView):
-    template_name = 'client/form_edu.html'
+    template_name = 'client/edit_forms/form_edu.html'
 
     def get(self, request, *args, **kwargs):
         client_instance = client_check(request.user)
-        response = defaultdict()
+        load_data = education_page_get(client_instance)['cl_edu']
 
-        load_data = load_education_page(client_instance)['cl_edu']
-        # load_data_edu = Education.objects.filter(client_edu=client_instance).values()
-        # edu_id = [e['id'] for e in load_data_edu]
-        # load_data_cert = [[c for c in Certificate.objects.filter(education_id=i).values()] for i in edu_id]
-        # print(load_data_cert[0])
-
-        response['client_img'] = load_client_img(client_instance)
-        response['edu_form'] = EducationFormSet(initial=load_data)
-        response['certificate'] = CertificateFormSet(initial=load_data)
-
+        response = {'client_img': load_client_img(client_instance),
+                    'edu_form': EducationFormSet(initial=load_data),
+                    # 'certificate': CertificateFormSet(initial=load_data),
+                    'certificate': CertificateForm(initial=load_data[0]),
+                    'sab_class_form': SabClassFormSet(initial=load_data),
+                    }
         return render(request, self.template_name, response)
 
     def post(self, request):
-        print("FormEducation.POST: %s" % request.POST)
         client_instance = client_check(request.user)
-        edu_inst = None
-        form_set_edu = EducationFormSet(request.POST)
-        form_set_cert = CertificateFormSet(request.POST, request.FILES)
-
-        if form_set_edu.is_valid():
-            print('FormSet_Edu - OK')
-            for f in form_set_edu:
-                f_items = f.cleaned_data.items()
-                print("edu_items: %s" % f_items)
-                if f_items:
-                    """ edu_inst - unsaved model instance!
-                    It gives you ability to attach data to the instance before saving to the DB! """
-                    edu_inst = f.save(commit=False)
-                    """ attach ForeignKey == Client instance """
-                    edu_inst.client_edu = client_instance
-                    """ Save Education instance """
-                    edu_inst.save()
-
-        if form_set_cert.is_valid():
-            print("FormSet_Cert - OK")
-            for c in form_set_cert:
-                c_items = c.cleaned_data.items()
-                print('cert_items: %s' % c_items)
-                if c_items:
-                    cert_inst = c.save(commit=False)
-                    """ attach ForeignKey == Education instance """
-                    cert_inst.education = edu_inst
-                    cert_inst.save()
-        else:
-            print("FormSet_Cert not Valid")
-
+        form_edu_post(client_instance, request)
         return redirect(to='/client/edit/form_edu')
-
-    # else:
-    #     response['edu_form'] = EducationFormSet()
-    #     response['certificate'] = CertificateFormSet()
-    #     response['inlineEduCert'] = inlineEduCert()
-
-   # return render(request, 'client/form_edu.html', response)
-
-
-def load_client_img(req):
-    """ Show Client Img in the Navigation Bar.
-    Img loaded from DB, if user do not have img - load default. """
-    try:
-        print("user: %s, id: %s" % (req, req.id))
-        client_img = Client.objects.get(user_client=req).img
-        if client_img:
-            logging.info("Client.img: %s" % client_img)
-            return "%s%s" % (MEDIA_URL, client_img)
-        else:
-            return '/media/user_1.png'
-    except Exception as ex:
-        logging.error("Exception in - load_client_img()\n %s" % ex)
-        return '/media/user_1.png'
-
 
 
 def checktask(request):
@@ -652,7 +381,7 @@ def chat_update(request):
     return JsonResponse(send2, safe=False)
 
 
-#Poland's views
+# Poland's views
 
 def vacancies_list(request, slug):
     resume = Resume.objects.get(slug__iexact=slug)
@@ -682,19 +411,20 @@ def resume_detail(request, slug):
     return render(request, 'client/client_resume_detail.html', context={'resume': resume})
 
 
-def accepted_vacancies(request, slug):########################
+def accepted_vacancies(request, slug):  ########################
     resume = Resume.objects.get(slug__iexact=slug)
     return render(request, 'client/client_accepted_vacancies.html', context={'resume': resume})
 
 
-def rejected_vacancies(request, slug):##############################
+def rejected_vacancies(request, slug):  ##############################
     resume = Resume.objects.get(slug__iexact=slug)
     return render(request, 'client/client_rejected_vacancies.html', context={'resume': resume})
 
 
-def accept_reject(request):#
+def accept_reject(request):  #
 
-    if request.GET['flag'] == 'accept' and Vacancy.objects.get(slug__iexact=request.GET['slug']).in_waiting_for_resume.all():
+    if request.GET['flag'] == 'accept' and Vacancy.objects.get(
+            slug__iexact=request.GET['slug']).in_waiting_for_resume.all():
         print(request.GET['slug'], 1)
         r = Vacancy.objects.get(slug__iexact=request.GET['slug']).in_waiting_for_resume.get()
         v = Vacancy.objects.get(slug__iexact=request.GET['slug'])
@@ -703,7 +433,8 @@ def accept_reject(request):#
         r.save()
         return HttpResponse('accept_server')
 
-    elif request.GET['flag'] == 'reject' and Vacancy.objects.get(slug__iexact=request.GET['slug']).in_waiting_for_resume.all():
+    elif request.GET['flag'] == 'reject' and Vacancy.objects.get(
+            slug__iexact=request.GET['slug']).in_waiting_for_resume.all():
         print(request.GET['slug'], 2)
         r = Vacancy.objects.get(slug__iexact=request.GET['slug']).in_waiting_for_resume.get()
         v = Vacancy.objects.get(slug__iexact=request.GET['slug'])
@@ -712,7 +443,8 @@ def accept_reject(request):#
         r.save()
         return HttpResponse('reject_server')
 
-    elif request.GET['flag'] == 'accept' and Vacancy.objects.get(slug__iexact=request.GET['slug']).reject_for_resume.all():
+    elif request.GET['flag'] == 'accept' and Vacancy.objects.get(
+            slug__iexact=request.GET['slug']).reject_for_resume.all():
         print(request.GET['slug'], 3)
         r = Vacancy.objects.get(slug__iexact=request.GET['slug']).reject_for_resume.get()
         v = Vacancy.objects.get(slug__iexact=request.GET['slug'])
@@ -721,7 +453,8 @@ def accept_reject(request):#
         r.save()
         return HttpResponse('accept_server')
 
-    elif request.GET['flag'] == 'reject' and Vacancy.objects.get(slug__iexact=request.GET['slug']).accept_for_resume.all():
+    elif request.GET['flag'] == 'reject' and Vacancy.objects.get(
+            slug__iexact=request.GET['slug']).accept_for_resume.all():
         print(request.GET['slug'], 4)
         r = Vacancy.objects.get(slug__iexact=request.GET['slug']).accept_for_resume.get()
         v = Vacancy.objects.get(slug__iexact=request.GET['slug'])
@@ -742,7 +475,7 @@ def settings_list(request):
     settings = Settings.objects.all()
     status = 1 if Settings.objects.get().tumbler_on_off == 'on' else 0
     print('status = ', status)
-    return render(request, 'client/settings.html', context={'settings': settings, 'status': status })
+    return render(request, 'client/settings.html', context={'settings': settings, 'status': status})
 
 
 #def settings_on_off(request):
@@ -767,6 +500,6 @@ def viewed(request):
             r.notification.clear()
         return HttpResponse('cleared')
 
-#End Poland's views
+# End Poland's views
